feat:make different file for read controller, changes in routes.js
