+++ backend/controllers/yogaPoses/createController.js
@@ -1,15 +1,5 @@
 const YogaPose = require("../../models/YogaPoseModel");
 
-// controller to get all yoga pose
-const getAllYogaPoses = async (req, res) => {
-  try {
-    const yogaPoses = await YogaPose.find();
-    res.json(yogaPoses);
-  } catch (error) {
-    res.status(500).json({ message: error.message });
-  }
-};
-
 // controller to create the new yoga pose
 
 const createYogaPose = async (req, res) => {
@@ -34,4 +24,4 @@ const createYogaPose = async (req, res) => {
   }
 };
 
-module.exports = { getAllYogaPoses, createYogaPose };
+module.exports = { createYogaPose };
